+++ apps/website/src/app/layout.tsx
@@ -2,7 +2,7 @@ import './global.css';
 
 import type { ReactNode } from 'react';
 
-import { ThemeProvider, ThemeScript } from '@vapor-ui/core';
+import { ThemeProvider } from '@vapor-ui/core';
 import { RootProvider } from 'fumadocs-ui/provider';
 import { Inter } from 'next/font/google';
 import Script from 'next/script';
@@ -17,11 +17,6 @@ export default function Layout({ children }: { children: ReactNode }) {
     return (
         <html lang="ko" className={inter.className} suppressHydrationWarning>
             <head>
-                <ThemeScript
-                    config={{
-                        storageKey: 'vapor-ui-theme',
-                    }}
-                />
                 <link rel="icon" href="/favicon.ico" sizes="any" />
                 <Script type="application/ld+json" id="vapor-ui-schema">
                     {JSON.stringify({
@@ -43,13 +38,9 @@ export default function Layout({ children }: { children: ReactNode }) {
                     }}
                     theme={{ enabled: false }}
                 >
-                    <ThemeProvider
-                        config={{
-                            storageKey: 'vapor-ui-theme',
-                        }}
-                    >
-                        {children}
-                    </ThemeProvider>
+                    {typeof window === 'undefined' ? (
+                        <ThemeProvider>{children}</ThemeProvider>
+                    ) : null}
                 </RootProvider>
             </body>
         </html>