+++ packages/core/src/components/field/index.ts
@@ -1,8 +1,1 @@
-export { Field, FieldRoot, FieldLabel, FieldDescription, FieldError, FieldSuccess } from './field';
-export type {
-    FieldRootProps,
-    FieldLabelProps,
-    FieldDescriptionProps,
-    FieldErrorProps,
-    FieldSuccessProps,
-} from './field';
+export * from './field';